+++ scene/animation/animation_blend_space_2d.cpp
@@ -46,7 +46,7 @@ Variant AnimationNodeBlendSpace2D::get_parameter_default_value(const StringName 
 	}
 
 	if (p_parameter == closest) {
-		return (int)-1;
+		return -1;
 	} else {
 		return Vector2();
 	}
